编译支持x64，但目前由于assimp不能够编译成x64，所以还是无法生成x64的工程

+++ KYEngine/Graphic/GraphicDef.h
@@ -563,8 +563,8 @@ namespace KY
 	{
 		const uint8 *pData;
 		//{@ valid for texture resource
-		size_t pitch;
-		size_t slicePitch;
+		uint32 pitch;
+		uint32 slicePitch;
 		//@}
 	};
